+++ example/helium_api_client_example.dart
@@ -6,7 +6,7 @@ void main() async {
   final address =
       '11yXFNu9Je9gvVfyuA8jS8z6faNcZW7kPwPApWYThTk416LGmU'; // acidic red mouse
   //'11cxkqa2PjpJ9YgY9qK3Njn4uSFu6dyK9xV8XE4ahFSqN1YN2db'; // tall plum griffin
-  var client = HeliumClient();
+  var client = HeliumBlockchainClient();
   var hotspotResp = await client.hotspots.getHotspotForAddress(address);
 
   print('Current:');