+++ CHANGELOG.md
@@ -1,12 +1,15 @@
-# 0.1.1
+# TBD  !Change this before committing!
 - Removed some private methods that accidentally escaped into the public API.
 - Removed dependency on dart:io to enable js support.
 - Updated README with basic usage information.
-
+- Renamed HeliumClient to HeliumBlockchainClient.
+- Renamed HeliumPoCReceiptsPathElement to HeliumPoCPathElement.
+- Renamed HeliumPoCReceiptsWitness to HeliumPoCWitness.
+- Renamed HeliumPoCReceiptsReceipt to HeliumPoCReceipt.
+- Documented more of the public API.
 
 # 0.1.0
 
 - Initial version.
 - Supports 3 APIs: Hotspots, Oracle Prices, and Transactions.
 - API method names and class names are beta and may be adjusted soon in a new major version.
-